requirements.txt
Django==3.1.5
djangorestframework==20.3.3
gunicorn
django-heroku